// athenas-frontend/src/enums/tipo-dependente-ir.enum.ts
export enum TipoDependenteIrEnum {
    CONJUGE = 1,
    COMPANHEIRO_UNIAO_ESTAVEL = 2,
    FILHO_FILHA_NAO_EMANCIPADO_21_ANOS = 3,
    FILHO_FILHA_ABSOLUTAMENTE_INCAPAZ = 4,
    PAI_MAE_COM_DEPENDENCIA_ECONOMICA = 5,
    IRMAO_NAO_EMANCIPADO_MENOR_21_ANO_COM_DEPENDENCIA_ECONOMICA_E_GUARDA_JUDICIAL = 6,
    IRMAO_ABSOLUTAMENTE_INCAPAZ = 7,
    ENTEADO_NAO_EMANCIPADO_21_ANOS = 8,
    ENTEADO_ABSOLUTAMENTE_INCAPAZ = 9,
    MENOR_TUTELADO_NAO_EMANCIPADO_MENOR_21_ANO_COM_DEPENDENCIA_ECONOMICA_OU_GUARDA_JUDICIAL = 10,
    MENOR_ABSOLUTAMENTE_INCAPAZ = 11,
    AVOS_COM_DEPENDENCIA_ECONOMICA = 12,
    BISAVOS_COM_DEPENDENCIA_ECONOMICA = 13,
    NETO_NETA_NAO_EMANCIPADO_MENOR_21_ANOS_COM_DEPENDENCIA_ECONOMICA_E_GUARDA_JUDICIAL = 14,
    BISNETO_BISNETA_NAO_EMANCIPADO_MENOR_21_ANOS_COM_DEPENDENCIA_ECONOMICA_E_GUARDA_JUDICIAL = 15,
    EX_CONJUGE = 16,
    FILHO_FILHA_OU_ENTEADO_ENTEADA_UNIVESITARIO_OU_CURSANDO_ESCOLA_TECNICA_2_GRAU_ATE_24_ANOS = 17,
    AGREGADOS_OUTROS = 18
}